Adding scorers to memory
