+++ pyrit/score/substring_scorer.py
@@ -2,6 +2,8 @@
 # Licensed under the MIT license.
 
 import asyncio
+from pyrit.memory.duckdb_memory import DuckDBMemory
+from pyrit.memory.memory_interface import MemoryInterface
 from pyrit.models.prompt_request_piece import PromptRequestPiece
 from pyrit.score import Score, Scorer
 
@@ -11,8 +13,9 @@ class SubStringScorer(Scorer):
     Scorer that checks if a given substring is present in the text.
     """
 
-    def __init__(self, *, substring: str, category: str = None) -> None:
-        super().__init__()
+    def __init__(self, *, substring: str, category: str = None, memory: MemoryInterface = None) -> None:
+        self._memory = memory if memory else DuckDBMemory()
+
         self._substring = substring
         self._category = category
         self.scorer_type = "true_false"
@@ -25,7 +28,7 @@ class SubStringScorer(Scorer):
 
         expected_output_substring_present = self._substring in request_response.converted_value
 
-        return [
+        score = [
             Score(
                 score_value=str(expected_output_substring_present),
                 score_value_description=None,
@@ -37,6 +40,9 @@ class SubStringScorer(Scorer):
                 prompt_request_response_id=request_response.id,
             )
         ]
+
+        self._memory.add_scores_to_memory(scores=score)
+        return score
 
     def validate(self, request_response: PromptRequestPiece):
         if request_response.converted_value_data_type != "text":
